Fix SQL error when adding dashboard condition
Some widgets and reports are filled with data that is generated by
multiple database queries. When the first query doesn't return any
matching data, we substitute a dummy query for the second SQL statement
so the final transaction result will be empty. The static dummy query
we had been using was flawed because it didn't look at the number of
query arguments that were going to be passed along with the query.
This would generate sql: expected x arguments, got y errors.
I updated the logic to generate the dummy query dynamically so it will
always reference all of the query arguments that are passed. I also
fixed several logger.Debug statement that were passing too many format
arguments which resulted in confusing MISSING %v messages in the log.