Fix netspace manager to limit available block size request
During an engineering review we determine that our goal with this
function is to provide reasonable default blocks of address space
that don't conflict with existing networks. To that end, I have
limited the networkSize for requests to a value in the 24-32 range.
Requests with a size outside that range will have the size adjusted
to 24. I renamed the address field in response to network to prevent
confusion since we are returning the network address of the space and
not the address of a host within it. Example JSON replies:

{"cidr":"10.49.99.0/24","netsize":24,"network":"10.49.99.0"}
{"cidr":"172.18.216.128/25","netsize":25,"network":"172.18.216.128"}