Fix to stop caching unique prepared statements
The UI generates some SQL queries that include timestamps and other
values inline as part of the query body. This results in complex and
unique SQL queries that will always be different. We don't want to
cache these prepared statements, as they will fill our prepared
statement cache endlessly. Fix was to stop caching UI queries for
SERIES and CATEGORIES_SERIES type reports, since these are the ones
that use the problematic syntax.