Rework the interface_stats and session_stats insert logic
Changing the session and interface stats to use a stored procedure
was fairly easy since we only ever perform INSERT statements on
those tables. Doing the same for the sessions table turns out to
be a much bigger exercise. Yes, I tried, and reverted back to
what we have here for now. This may cause a merge conflict with
master as I had to include the is_wan stuff we added to master
for MFW-1012 as part of the logic here. I also need to finish
the logic to format the cloud event for WAN interface stats so
still a work in progress.