pinger: Use ListenIP/IPConn
The sockets we use to ping for active latency testing need to
be bound to particular interfaces.  This isn't possible with via
the icmp.PacketConn interface.  So, switch to calling net.ListenIP
in order to get a net.IPConn connection, which allows us to get a
handle on the underlying file descriptor, which allows us to use a
syscall to bind it to a particular interface.

MFW-1234