Finish initial rework of traffic prediction service
The CoreDNS plugin approach turned out to be a bad idea. Their
framework is very DNS specific which makes it difficult to leverage
for passing the more detailed and free-form query data we need.
So I wrote a new predictor daemon instead. It's a small, multithreaded
pure C daemon that receives UDP queries, fetches the relevant info
from a mysql database, and sends the response to the client.