Rework getAvailableAddressSpace function
Reworked the getAvailableAddressSpace function with the same logic
used in the NGFW. IPv6 addresses are still generated randomly in the
ULA block defined in RFC4193. For IPv4 we start at the beginning of
the reserved blocks defined in RFC1918 looking until we find one
that is not use. Subsequent calls will continue searching at the
network following the last one we returned.